added graphics library

+++ src/stnd.h
@@ -2,7 +2,7 @@
 #include <time.h>
 
 
-void delay(int number_of_seconds)
+void delay(float number_of_seconds)
 {
     // Converting time into milli_seconds
     int milli_seconds = 1000 * number_of_seconds;
